Resend Request - Wrote Prototypical Acceptance Test
In this commit, I wrote the acceptance test to prototype the solution I
want to implement for the published event resend request.  I also
attempted a solution that failed, though I learned a lot about what the
impact to the event system will be.  I now have a solution in mind, but
I'll need to refactor a number of things to get it working.