Resend Request - Refactor Subscriber Structure
In this commit, in preparation for implementing the resend published
event request feature, I refactored the code structure for the
Subscriber.  Essentially, I extended and corrected what was already in
place.  Internally, the event system uses the SubscriberInternalImp
subscriber implementation for all activity.  It is a separate
implementation hierarchy from the externally published Subscriber base
class that ONLY implements the SubscriberPublished interface.  This
solution better insulates the event system in case of a wonky external
Subscriber class implementation.