Resend Request - Refactor In Preparation To Fix Broken Acceptance Test
In this commit, I refactored some code in the channel implementation to
make it easier to implement the fix for the broken acceptance test.  I
just pulled all the logic related to updating the state of the published
events out of the publish / unpublish methods and into four private
methods.  It will be these private methods that will be modified to
include some additional data structures to track the order of published
events (a List).