JavaDoc - Updated For Delete Support Feature
In this commit, I updated the existing JavaDoc to include the
documentation for the new delete feature that was implemented.  I also
did some various clean up so that the JavaDoc follows the same
conventions as the the JDK8 JavaDoc.  Finally, I also cleaned up the
imports in EventCoreAccpetanceTests.java.