Refactor - Start Refactor of EventFactory Implementation
In this commit, I started refactoring the EventFactory implementation so
that it is more self contained and better structured.  I renamed some of
the classes and methods to refer to the starting point of the
EventFactory implementation as the "head" as opposed to the "root".
This makes more sense since the implementation is more similar to a
linked list as opposed to a tree.  I also added a base class to
implement the default behaviour of delegating to the next layer in the
EventFactory so that only those methods that actually do something extra
need to be implemented by the head, parameter validator, cacher, and
creator layers.