Refactor - Add EventFactory Tail Implementation
In this commit, I implemented an EventFactory Tail class that just
throws an exception if ANY EventFactory method is reached.  This tail
implementation was added as the next EventFactory after the creator.
Any existing creator methods that were throwing exceptions were removed
from the creator implementation so that throwing the exceptions would be
delegated to the tail EventFactory.  That way, the responsibility of
each layer of the EventFactory is even more evident.