Refactor - Move Parameter Validation To Correct Class
In this commit, I moved some more parameter validation to the parameter
validation class.  The problem was that the reference to the instance
cache was in the held by the cache EventFactory layer.  But this didn't
really make sense because not only is the instance cache required for
parameter validation, it is also referenced by other classes within the
Event System.  So, I moved the reference into the Head EventFactory
layer.