Refactor - Rename InstanceCache and ChannelCache
In this commit, I renamed InstanceCache and ChannelCache to
InstanceCacheInternal and ChannelCacheInternal to reflect that these
classes are implementation artifacts that are not accessible outside the
Java package.  I'm pretty sure this practice goes against Clean Coding
best practices but it does help me keep track of what classes and
methods are externally accessible.  Maybe with more practice I won't
need to do this?  But then, how would I differentiate between internally
and externally defined interfaces?