Refactor - Rename Another Reference To ChannelCache
In this commit, I renamed another reference to ChannelCache that I
missed in the previous commit.  I also prototyped changing the structure
of the code to have definitions of all the methods in the published
EventFactory interface that deal with only the Internal classes, such as
ChannelInternal, EventDescriptionInternal, etc., since it is these types
that are actually referenced by all the implementations, internally.
The objective is to get rid of all the casting that is done in a number
of classes when actual processing occurs.  The implementation I
attempted failed, though since I thought it would be best to cast early,
in the head EventFactoryInternal implementation.  I was wrong, though
since a number of the parameter validation methods would need to be
executed before we could safely cast from the published to the internal
event types.  Perhaps another link in the chain could be added after the
validator and before the cacher, but I am not certain that shifting from
published to internal in such a way would actually make the code easier
to understand.  I'll try it, though.