Refactor - Channel Map To Internal
In this commit, I implemented a similar solution for Channel that I did
for both EventDescription and Publisher so that work related to the
Channel type is done with respect to the ChannelInternal type.  Note
that Subscriber uses a different strategy for mapping between the
published Subscriber type and the SubscriberInternal type used in the
system.  This is due to the fact that Subscriber is expected to be an
external implementation so it is wrapped with a SubscriberInternal
implementation that quarantines the external Subscriber implementation
from the rest of the event system.  This would be a good solution to use
in general except that part of the Subscriber implementation leaks out
in the Subscriber published interface.  If I fix that issue, then maybe
it would make sense to use it throughout as a more general solution for
safely separating out the external published type from the type used
internally.