Fix Eclipse Build Failure - Upgrade Gradle
In this commit, I am fixing the issues with the Event project Eclipse
build.  The issue was that I could not build the Event project because
it was not loading the dependencies for both build and test for some
reason.  I assumed that the problem was that the gradle version being
used by Eclipse was the issue.  So, here is what I did to correct the
problem:
-Deleted the local copy of Event and recloned it from GitHub.
-Determined that the Gradle Wrapper was not compatible with the
installed Java version.
-Downloaded latest version of Gradle.
-Updated the Gradle Wrapper to the newest version of Gradle.
-Fixed build issue because an 'implementation' dependency was missing.
-Fixed repositories since jcenter is now read-only so changed to maven
central.
-Added the 'eclipse' plugin to the gradle build.
-Changed the plugin declarations to add the newer, preferred syntax.
-Tested building the project from the commandline.  It worked.
-Tested importing the project into Eclipse and building it.  It worked.

I think that's it!